fix: load API on detail page
