+++ composables/webapis.ts
@@ -36,3 +36,21 @@ export const useWebApiStatus = (api: WebAPI, available?: boolean) => {
     return undefined
   })
 }
+
+export const useLoadWebApis = (webApis?: WebAPI[]) => {
+  if (navigator) {
+    const webApiStatuses: Ref<{ [key: keyof typeof webApiData]: boolean }> = useState('webApiStatuses', () => ({}))
+    webApis = webApis || useWebApiList().value
+    webApis.forEach((webApi) => {
+      if (!webApiStatuses.value[webApi.id]) {
+        const check = webApi.check || defaultWebApiCheck
+        if (check.constructor.name === 'AsyncFunction') {
+          (check(webApi) as Promise<boolean>)
+            .then((available: boolean) => webApiStatuses.value[webApi.id] = available)
+        } else {
+          webApiStatuses.value[webApi.id] = check(webApi) as boolean
+        }
+      }
+    })
+  }
+}
